move offhours to filters pkg, update readme
